# docs/swagger.yaml
basePath: /api/v1
definitions:
  httpadapter.AuthResponse:
    properties:
      token:
        type: string
    type: object
  httpadapter.CreateItemRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      priority:
        maximum: 5
        minimum: 1
        type: integer
      url:
        type: string
    required:
    - name
    - priority
    type: object
  httpadapter.CreateWishlistRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      event_date:
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - event_date
    - title
    type: object
  httpadapter.ItemResponse:
    properties:
      description:
        type: string
      id:
        type: string
      is_reserved:
        type: boolean
      name:
        type: string
      priority:
        type: integer
      reserved_at:
        type: string
      url:
        type: string
      wishlist_id:
        type: string
    type: object
  httpadapter.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  httpadapter.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  httpadapter.UpdateItemRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      priority:
        maximum: 5
        minimum: 1
        type: integer
      url:
        type: string
    type: object
  httpadapter.UpdateWishlistRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      event_date:
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
    type: object
  httpadapter.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
    type: object
  httpadapter.WishlistResponse:
    properties:
      description:
        type: string
      event_date:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/httpadapter.ItemResponse'
        type: array
      public_token:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API для управления вишлистами подарков
  title: Wishlist API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для входа
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpadapter.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpadapter.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вход
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для регистрации
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpadapter.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpadapter.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация
      tags:
      - auth
  /public/{token}:
    get:
      parameters:
      - description: Публичный токен вишлиста
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpadapter.WishlistResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить публичный вишлист
      tags:
      - public
  /public/{token}/items/{itemID}/reserve:
    post:
      parameters:
      - description: Публичный токен вишлиста
        in: path
        name: token
        required: true
        type: string
      - description: ID подарка
        in: path
        name: itemID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Забронировать подарок
      tags:
      - public
  /wishlists:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/httpadapter.WishlistResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Список вишлистов
      tags:
      - wishlists
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные вишлиста
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpadapter.CreateWishlistRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpadapter.WishlistResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создать вишлист
      tags:
      - wishlists
  /wishlists/{id}:
    delete:
      parameters:
      - description: ID вишлиста
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить вишлист
      tags:
      - wishlists
    get:
      parameters:
      - description: ID вишлиста
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpadapter.WishlistResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить вишлист
      tags:
      - wishlists
    put:
      consumes:
      - application/json
      parameters:
      - description: ID вишлиста
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpadapter.UpdateWishlistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpadapter.WishlistResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить вишлист
      tags:
      - wishlists
  /wishlists/{id}/items:
    get:
      parameters:
      - description: ID вишлиста
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/httpadapter.ItemResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Список подарков
      tags:
      - items
    post:
      consumes:
      - application/json
      parameters:
      - description: ID вишлиста
        in: path
        name: id
        required: true
        type: string
      - description: Данные подарка
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpadapter.CreateItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpadapter.ItemResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создать подарок
      tags:
      - items
  /wishlists/{id}/items/{itemID}:
    delete:
      parameters:
      - description: ID вишлиста
        in: path
        name: id
        required: true
        type: string
      - description: ID подарка
        in: path
        name: itemID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить подарок
      tags:
      - items
    put:
      consumes:
      - application/json
      parameters:
      - description: ID вишлиста
        in: path
        name: id
        required: true
        type: string
      - description: ID подарка
        in: path
        name: itemID
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpadapter.UpdateItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpadapter.ItemResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить подарок
      tags:
      - items
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"